chore: Move `LinkMapKey` to `miden-objects`
